Add spring security + jaas

+++ lab1/src/main/kotlin/com/itmo/blss/api/response/TrainDto.kt
@@ -2,6 +2,6 @@ package com.itmo.blss.api.response
 
 data class TrainDto(
     val trainId: Long,
-    val trainNum: Int
+    val trainNum: Long
 ){
 }
